refactor: model 패키지 제거 및 Card domain 패키지 리팩토링
