+++ README.md
@@ -4,7 +4,7 @@
 - [x] Deck 은 52장의 중복 없는 카드로 이루어진다
 - [x] Deck 에서 draw 하면 Deck 의 카드 수는 1 감소해야 한다
 - [x] Deck 에서 draw 한 카드는 Deck 에서 사라져야 한다
-- [ ] Hand 는 초기에 2 장의 카드로 이루어져야 한다
+- [x] Hand 는 초기에 2 장의 카드로 이루어져야 한다
 - [x] Hand 의 sum 은 정확한 값을 반환하야 한다
 - [x] Hand 의 sum 이 21이면 blackjack 이다
 - [x] Hand 의 sum 이 21을 초과하면 bust 이다
