+++ src/main/kotlin/blackjack/domain/card/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.domain.card
 
 enum class Suit {
     Spade, Diamond, Heart, Clover
