+++ src/test/kotlin/blackjack/domain/cards/DeckTest.kt
@@ -1,8 +1,8 @@
-package blackjack.domain.card
+package blackjack.domain.cards
 
-import blackjack.model.Card
-import blackjack.model.Character
-import blackjack.model.Suit
+import blackjack.domain.card.Card
+import blackjack.domain.card.Character
+import blackjack.domain.card.Suit
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import org.assertj.core.api.Assertions
