+++ src/test/kotlin/blackjack/domain/player/PlayerTest.kt
@@ -1,9 +1,9 @@
 package blackjack.domain.player
 
-import blackjack.domain.card.HandCards
-import blackjack.model.Card
-import blackjack.model.Character
-import blackjack.model.Suit
+import blackjack.domain.card.Card
+import blackjack.domain.card.Character
+import blackjack.domain.card.Suit
+import blackjack.domain.cards.HandCards
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import org.assertj.core.api.Assertions
